Updated GLA Template to 1.31
Moved Jquery.js & cookies.js out of folder in GLA template, because
apparently I missed that and we've been copying them every time. The
GLAs aren't updated with this change because it has a big potential to
break everything and that's more important than space at this point. But
I updated the template so when we make new ones it'll be that way.